exercises/collector/solution/requirements.txt
Faker==28.4.1
Flask==3.0.3
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
requests==2.32.3